+++ src/Application/Rule/Rules/ConditionRule.php
@@ -3,12 +3,12 @@
 namespace ArtARTs36\MergeRequestLinter\Application\Rule\Rules;
 
 use ArtARTs36\MergeRequestLinter\Domain\Condition\ConditionOperator;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\Counter;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\NullCounter;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDecorator;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Counter;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\NullCounter;
 
 class ConditionRule implements RuleDecorator
 {
