+++ src/Infrastructure/RequestFetcher/CiRequestFetcher.php
@@ -3,12 +3,12 @@
 namespace ArtARTs36\MergeRequestLinter\Infrastructure\RequestFetcher;
 
 use ArtARTs36\MergeRequestLinter\Domain\CI\CurrentlyNotMergeRequestException;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricManager;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricSubject;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\StringMetric;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequestFetcher;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\CI\CiSystemFactory;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricManager;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricSubject;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\StringMetric;
 
 class CiRequestFetcher implements MergeRequestFetcher
 {
