+++ src/Infrastructure/Rule/Factories/ConditionRuleFactory.php
@@ -3,12 +3,12 @@
 namespace ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Factories;
 
 use ArtARTs36\MergeRequestLinter\Application\Rule\Rules\ConditionRule;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\Counter;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricManager;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricSubject;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Condition\OperatorResolver;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Counter;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricManager;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricSubject;
 
 /**
  * @phpstan-import-type Conditions from OperatorResolver
@@ -24,7 +24,7 @@ class ConditionRuleFactory
 
     public static function new(OperatorResolver $operatorResolver, MetricManager $metrics): self
     {
-        $counter = new MemoryCounter();
+        $counter = new IncCounter();
 
         $metrics->add(new MetricSubject('linter_skipped_rules', '[Linter] Skipped rules'), $counter);
