+++ src/Shared/Metrics/Manager/NullMetricManager.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager;
+namespace ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager;
 
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\Metric;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricManager;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricSubject;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Metric;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricManager;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricSubject;
 
 /**
  * @codeCoverageIgnore
