+++ src/Shared/Time/Duration.php
@@ -2,7 +2,9 @@
 
 namespace ArtARTs36\MergeRequestLinter\Shared\Time;
 
-class Duration
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Metric;
+
+class Duration implements Metric
 {
     public function __construct(
         public readonly float $seconds,
@@ -13,5 +15,10 @@ class Duration
     public function __toString(): string
     {
         return "$this->seconds" . 's';
+    }
+
+    public function getMetricValue(): string
+    {
+        return (string) $this;
     }
 }
