+++ tests/Unit/Domain/Metrics/MemoryCounterTest.php
@@ -2,31 +2,31 @@
 
 namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Domain\Metrics;
 
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class MemoryCounterTest extends TestCase
 {
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::create
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::getMetricValue
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::create
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::getMetricValue
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::__construct
      */
     public function testCreate(): void
     {
-        $counter = MemoryCounter::create([0, 1, 2]);
+        $counter = IncCounter::create([0, 1, 2]);
 
         self::assertEquals('3', $counter->getMetricValue());
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::inc
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::getMetricValue
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::inc
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::getMetricValue
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter::__construct
      */
     public function testInc(): void
     {
-        $counter = new MemoryCounter(2);
+        $counter = new IncCounter(2);
 
         $counter->inc();
