+++ tests/Unit/Domain/Metrics/MetricProxyTest.php
@@ -2,16 +2,16 @@
 
 namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Domain\Metrics;
 
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\Metric;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricProxy;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Metric;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricProxy;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class MetricProxyTest extends TestCase
 {
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricProxy::getMetricValue
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricProxy::retrieve
-     * @covers \ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricProxy::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricProxy::getMetricValue
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricProxy::retrieve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricProxy::__construct
      */
     public function testGetMetricValue(): void
     {
