+++ tests/Unit/Infrastructure/Metrics/MemoryMetricManagerTest.php
@@ -2,18 +2,18 @@
 
 namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Metrics;
 
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MemoryCounter;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\MetricSubject;
-use ArtARTs36\MergeRequestLinter\Domain\Metrics\Record;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager\MemoryMetricManager;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager\MemoryMetricManager;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\IncCounter;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\MetricSubject;
+use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Record;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\QueueClock;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class MemoryMetricManagerTest extends TestCase
 {
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager\MemoryMetricManager::add
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager\MemoryMetricManager::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager\MemoryMetricManager::add
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager\MemoryMetricManager::__construct
      */
     public function testAdd(): void
     {
@@ -21,7 +21,7 @@ final class MemoryMetricManagerTest extends TestCase
 
         self::assertCount(0, $manager->describe());
 
-        $manager->add(new MetricSubject('', ''), new MemoryCounter());
+        $manager->add(new MetricSubject('', ''), new IncCounter());
 
         self::assertCount(1, $manager->describe());
     }
@@ -31,7 +31,7 @@ final class MemoryMetricManagerTest extends TestCase
         return [
             [
                 [
-                    [$subject1 = new MetricSubject('k', 'n'), $metric1 = new MemoryCounter()],
+                    [$subject1 = new MetricSubject('k', 'n'), $metric1 = new IncCounter()],
                 ],
                 [
                     new Record($subject1, $metric1, new \DateTimeImmutable()),
@@ -41,8 +41,8 @@ final class MemoryMetricManagerTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager\MemoryMetricManager::describe
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Metrics\Manager\MemoryMetricManager::__construct
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager\MemoryMetricManager::describe
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Metrics\Manager\MemoryMetricManager::__construct
      * @dataProvider providerForTestDescribe
      */
     public function testDescribe(array $adds, array $expected): void
